test: enhance encryption tests (concrete use cases, more coverage)
